Fix prometheusHandler for test
- could not spwan multiples instances of the app with the function spawn_app for the intergration test because PrometheusBuilder::new install the recorder globally
- refactoring to be able to pass the prometheusHandler from the main function to be able to use OnceCell in the tests to init only once the prometheusHandler

+++ src/config.rs
@@ -39,7 +39,7 @@ pub struct Postgres {
     pub db_password: Secret<String>,
     pub db_host: Secret<String>,
     pub db_name: Secret<String>,
-    pub db_port: i64,
+    pub db_port: u16,
 }
 
 pub struct Crypt {
